Updated Observations for Logging DFD
Feel free to add on to or change this portion of the document. Wanted to discuss the design flaw of over-utilizing the Panda safety package, denial of service attacks, and physical access woes. While I feel this can properly reflect the analyses returned from the MFMT, it may need some extra additions to bring it up to spec. 

If any changes are made, please add the 7 digit commit number to the project card.